tests: reorganize and update files fixtures
